requirements.txt
opencv-python==4.8.1
numpy==1.24.3
pyautogui==0.9.54
python-dotenv==1.0.0
